Made automation parameters floats, and also fixed a potential bug in AutomationCurve.evaluate

+++ Sources/AudioKit/Nodes/Mixing/AKFader.swift
@@ -122,8 +122,8 @@ public class AKFader: AKNode, AKToggleable, AKComponent {
 
     /// Convenience function for adding a pair of points for both left and right addresses
     public func addAutomationPoint(value: AUValue,
-                                   at startTime: Double,
-                                   rampDuration: Double = 0,
+                                   at startTime: Float,
+                                   rampDuration: Float = 0,
                                    taper taperValue: Float = 1,
                                    skew skewValue: Float = 0) {
         let point = AKParameterAutomationPoint(targetValue: value,
